requirements.txt
# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
requests>=2.31.0
sqlalchemy>=2.0.0

# Authentication
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2  # Fixed version for passlib compatibility
PyJWT>=2.8.0
email-validator>=2.1.0

# ML & AI
pandas>=2.1.0
numpy>=1.24.0
scikit-learn==1.7.0  # Pinned version for model compatibility
joblib>=1.3.0
shap>=0.43.0  # SHAP explanations for ML models

# Frontend
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
streamlit-extras>=0.3.0
streamlit-option-menu>=0.3.12  # Modern navigation menu
streamlit-card>=0.0.4  # Card components
hydralit-components>=1.0.10  # Additional modern UI components

# Data Processing
tqdm>=4.65.0

# Deployment
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
httpx>=0.24.0

# Utilities
typing-extensions>=4.7.0
Pillow>=10.0.0  # For image handling

# Document processing
PyPDF2>=3.0.0  # PDF text extraction
pdfplumber>=0.10.0  # Alternative PDF extraction (better for complex PDFs)
python-docx>=1.0.0  # DOCX document processing
pytesseract>=0.3.10  # OCR for images (requires Tesseract installed on system)
python-multipart>=0.0.6  # For file upload handling in FastAPI

# LLM integration (optional - used for risk analysis if API key is available)
google-generativeai>=0.3.0  # Google's Gemini API client
groq>=0.4.0  # GROQ API client for fast inference (recommended)
